Implements  RAII for the creation and deletion of the wazuh-agent.lock file

+++ src/agent/src/unix/unix_daemon.cpp
@@ -6,8 +6,10 @@
 
 #include <cerrno>
 #include <filesystem>
+#include <fmt/format.h>
 #include <fstream>
 #include <sys/file.h>
+#include <unistd.h>
 
 #if defined(__linux__)
 #include <unistd.h>
@@ -23,22 +25,24 @@ namespace unix_daemon
 {
     LockFileHandler::LockFileHandler(std::string lockFilePath)
         : m_lockFilePath(std::move(lockFilePath))
+        , m_errno(0)
         , m_lockFileCreated(createLockFile())
     {
     }
 
-    bool LockFileHandler::removeLockFile() const
+    void LockFileHandler::removeLockFile() const
     {
+        if (!m_lockFileCreated)
+            return;
+
         const std::string filePath = fmt::format("{}/wazuh-agent.lock", m_lockFilePath);
         try
         {
             std::filesystem::remove(filePath);
-            return true;
         }
         catch (const std::filesystem::filesystem_error& e)
         {
             LogError("Error removing file: {}", e.what());
-            return false;
         }
     }
 
@@ -75,13 +79,15 @@ namespace unix_daemon
         int fd = open(filename.c_str(), O_WRONLY | O_CREAT | O_TRUNC, 0644);
         if (fd == -1)
         {
-            LogError("Unable to open lock file: {}. Error: {} ({})", filename.c_str(), errno, std::strerror(errno));
+            m_errno = errno;
+            LogError("Unable to open lock file: {}. Error: {} ({})", filename.c_str(), m_errno, std::strerror(m_errno));
             return false;
         }
 
         if (flock(fd, LOCK_EX | LOCK_NB) == -1)
         {
-            LogDebug("Unable to lock lock file: {}. Error: {} ({})", filename.c_str(), errno, std::strerror(errno));
+            m_errno = errno;
+            LogDebug("Unable to lock lock file: {}. Error: {} ({})", filename.c_str(), m_errno, std::strerror(m_errno));
             close(fd);
             return false;
         }
@@ -109,10 +115,17 @@ namespace unix_daemon
 
         if (!lockFileHandler.isLockFileCreated())
         {
-            return "running";
+            if (lockFileHandler.getErrno() == EAGAIN)
+            {
+                return "running";
+            }
+            else
+            {
+                return fmt::format(
+                    "Error: {} ({})", lockFileHandler.getErrno(), std::strerror(lockFileHandler.getErrno()));
+            }
         }
 
-        lockFileHandler.removeLockFile();
         return "stopped";
     }
 } // namespace unix_daemon
